audit: Increase maximum number of names logged per syscall to 30
Increase maximum number of names logged per syscall to 30, as the
current value of 20 can be too small.  This was enough for Rik Theys
who reported #631799.

There is a better but bigger fix upstream, commit 5195d8e217a7
('audit: dynamically allocate audit_names when not enough space is in
the names array').