alx: remove atl1c devices
When alx was first reviewed, upstream demanded that any changes to
support for the currently supported AR8131/8132/8151/8152 chips be
made as a series of patches, not by deprecating or removing atl1c in a
monolithic patch.  Therefore:

Luis R. Rodriguez <mcgrof@frijolero.org> wrote:
> The alx driver is to only support the AR8161 and AR8162
> devices. The older devices are supported through atl1c.

I've also changed the Kconfig help text to reflect this.