aufs: possible bugfix, barrier for i_nlink
Set i_nlink carefully since its value is tested in au_new_inode(). If
the value becomes 0, aufs produces an incorrect warning about UDBA.
Generally i_nlink is set with i_mutex (or other locks) held. But for
aufs virtual inode, it is set without i_mutex in very special case
(copyup with a different name for link(2)).
Prasad Koya reported this phenomenon in aufs2.2. aufs3.0 already changed
the assignment to be careful. This commit is just to be more careful.

Reported-by: Prasad Koya <prasad.koya@gmail.com>
Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>