Correct SATA SERDES config for 24 MHz ref clock
SATA did not work because the SERDES PHY configuration was based on a 48
MHz ref clock. Change the configuration values to work with a 24 MHz ref
clock and Rev A1 silicon.

This will break compatibility with the EVM board for now (It has a 48
MHz clock and Rev A0 silicon). We expect to import a more recent SDK
from Mindspeed soon which probably works with a 24 and 48 MHz ref clock
as well as A0 and A1 silicon.

Change-Id: I39db0de22b830b5656459d2e1442b8639e77f923