Add dedicated board file w/ config for Optimus
The design of Optimus is different from the C2K eval board. The board
file describes the details that are relevant for the kernel like, for
example, which PHY is connect to which RMGII interface. Let us keep the
board file for the eval board pristine and create a dedicated copy for
Optimus which we will change in subsequent commits.

Change-Id: Ic4bf89a506f6e67b3d16806b2a0fb3dd6f0c7d12