SPI interface settings for optimus boards.
This enables the MoCA interface, basically, since that's the only thing on
the SPI bus on this device.

We also take the MoCA device out of reset.  Note that system reset already
apparently puts the device back into reset (the reset GPIO pin gets set to
input by default, and a pullup resistor on the board pulls it low, ie.  back
into reset), so we don't need to do actually toggle it here.

Change-Id: I394a96718a5f3fc9eecc7b1a38c36a1efd66bfd6