i2c-comcerto: Set i2c bus speed to 400 kHz instead of 100 kHz.
This works for our two devices (at24c512 and lm96063).  So does 1 MHz, but
that is apparently too fast to be legal on a "normal" i2c bus.  1 MHz is
supported by our flash chip according to its data sheet, but not officially
supported by lm96063, so let's be safe about it.  Unfortunately there's no
way in the driver to set a different speed for each device, unlike SPI.

With this change, you can read the at24c512 eeprom in about 1.5 seconds,
rather than 6.5 seconds.

Change-Id: I93587de7d1b785810ad1332d738d5c38ab9b4fcc