nand_base: with >= 224-byte ECC, you can tolerate a few bitflips.
UBI in particular was throwing away entire eraseblocks just because of a
single bitflip.  This is inappropriate in modern super-cheapo MLC NAND,
which can have a couple of bitflips in >50% of pages.  Instead, pick a
threshold that's still well under the maximum (24*4 bits per page) but much
more than the errors on a typical page.  If we flip fewer than that many
bits, we don't report the fact that ECC fixed stuff up, because UBI doesn't
need to know.

Change-Id: I82177298e7256f7672fd8701fcd63e6498310567