Disable NAND bad block table (bbt) and allow erasing bad blocks.
The bbt is only a speed optimization; bad block information is stored in the
OOB of each block already.  It's important for the bbt to be synchronized
between Linux and the bootloader, and we currently have no confidence that
this is the case, so just turn it off for now.  We can re-enable it later
when we have the luxury of lots of time to spend on testing it.

Bugs in earlier kernels ended up marking zillions of blocks as bad even
though they were fine (or mostly fine, within a reasonable range of ECC
bitflip correction).  This is unfixable if we can't erase bad blocks.  So
remove the (rather unnecessary anyway) hardcoded protection against that.
You can erase bad blocks, to give them another chance at life, using
"flash_erase -N".

Change-Id: Ifde40877534499fa889482cac4ea927e9a45850f