board-optimus: allow bmoca devices on either of the 2 SPI buses.
On the original optimus prototype boards, it was on the "slow" SPI bus. On
the new boards, it's on the "fast" SPI bus. The driver can fairly easily
detect whether a given device really exists or not, so let's just advertise
both and let it reject at runtime the ones that aren't there.

Change-Id: I9b025b892055f1d14110c71c254ec3cac74f2f60