Enable CONFIG_COMCERTO_CUSTOM_SKB_LAYOUT
This config option is necessary to benefit from Mindspeed's changes for
improved wifi performance.

As a side effect, this enables ZONE_DMA which in turn moves TEXT_OFFSET
from 0x00008000 to 0x04008000. This requires a change to our version of
barebox because it previously loaded the initrd also in the 0x04008000
range.

Change-Id: Idcea87a3a84718cbc39b2d81f56d9989d7dfec64