PCI/MSI: Export MSI mode using attributes, not kobjects
The PCI MSI sysfs code is a mess with kobjects for things that don't really
need to be kobjects.  This patch creates attributes dynamically for the MSI
interrupts instead of using kobjects.

Note, this removes a directory from sysfs.  Old MSI kobjects:

  pci_device
     └── msi_irqs
         └── 40
             └── mode

New MSI attributes:

  pci_device
     └── msi_irqs
         └── 40

As there was only one file "mode" with the kobject model, the interrupt
number is now a file that returns the "mode" of the interrupt (msi vs.
msix).

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Neil Horman <nhorman@tuxdriver.com>
[bwh: Adapted for Debian wheezy:
 - Adjusted to apply without commit 9738abedd672
   ('PCI: Make local functions/structs static')
 - Use sysfs_{create,remove}_group() instead of sysfs_{create,remove}_groups()
   (which don't exist in 3.2); change type and name of
   pci_device::msi_irq_groups accordingly
 - Move pci_device::msi_irq_group to end of the structure and hide it from
   genksyms]