Break infinite loop when modprobing nf_conntrack-
The kernel depends on a user mode helper to load kernel modules on
demand. The default helper is /sbin/modprobe, but this can be changed by
writing to /proc/sys/kernel/modprobe. If user mode writes an empty
string to this proc file, the kernel no longer spawns a helper process,
but just assumes the successful loading of the requested module. This
used to cause an infinite loop in nf_ct_l3proto_try_module_get().

Change-Id: Id60c6e697a5b4e8dd60d36117b33529cbc949ab1