gfrg200: change config settings to use wireless from linux-backports.
Disable mac80211 and cfg80211 in the kernel, because linux-backports
provides newer ones.  Enable a few encryption/checksum algorithms, because
linux-backports depends on them for some drivers.

We do want the WEXT stuff for now so that iwconfig/iwlist continue to work,
although we probably should switch entirely over to the new-style 'iw'
command, which doesn't need WEXT.

Change-Id: Ic01ce336f99cd038f148924c54bcac233437b72f