gfrg200: enable IGMP snooping.
This is already used on gfhd100/gfms100 devices.

Without this, multicast subscriptions on one port tend to get blasted to all
the ports, which can be disastrous on MoCA and wifi.  I don't know how this
interacts with mindspeed's hardware accelerated forwarding, but it seems
prudent to turn it for now since our briding is all in software at the
moment.

Change-Id: Id5bc798c01cc5408d2bb7e9ecac56d6b8b4d85ea