comcerto_nand: don't lie and say there were zero bitflips corrected.
The error correction function is expected to return the actual
number of recoverable ECC errors.  This happens when using software ECC, or
even in this function in the branch that does hamming correction.  But when
doing 24_HW_BCH, it always returns zero.  This prevents
ubifs from remapping almost-bad blocks.

Probably the reason they did this is that in this version of the kernel, the
default behaviour is to explode if there are nonzero recoverable errors.
That was the reason I added commit 868110c95, to add a higher
threshold depending which error correction mode is in use.  (Later upstream
kernels do the same thing.)  I think my way is safer than just pretending
there were no bit flips.  At the very least, this will give us some log
messages as people's flash is starting to go bad.

Change-Id: Ifbf301cf6c2d12343fb7af3731088ec5e0fe7b32