Add FALLOC_FL_NO_HIDE_STALE
Port of http://www.spinics.net/lists/linux-ext4/msg32702.html
to the Mindspeed 3.2.26 kernel.

HOWEVER: we allow KEEP_SIZE and NO_HIDE_STALE to be used at
the same time, as we need to know the actual size of the
video not the fallocated size. From the link above:

"Also note that I restrict the combination of NO_HIDE_STALE && KEEP_SIZE
since it causes e2fsck to complain --- and if you're trying to avoid fs
metadata I/O, you want to avoid the extra i_size update anyway, so it's
not worth trying to make this work w/o causing e2fsck complaints."

We intend to modify e2fsck in our tree to know about this combination
and not complain.

Change-Id: Ic50ab699f32b6d383ef2062beebc8b238ee7e391