Mindspeed kernel: add new UDP socket option to report RX queue level
(this is the exact same change already applied to the bruno kernel via
https://gfiber-internal-review.googlesource.com/54693)

This adds a new socket option “SO_RXQ_ALLOC” that enables providing
the RX queue buffer allocation as ancillary data from the recvmsg()
system call. The value reported is a byte number and together with the
RX queue size (obtained via getsockopt(SO_RCVBUF) can be used to
calculate a percentage value.

Change-Id: I4fbadc6b9ef7549cb526cf78b7e432d8054370f5