SpaceCast: Define Ethernet ifs in board config file
This change adds the SpaceCast board specific Change
to board-optimus.c. In SpaceCast, only PHY 1 GEMAC 1
is supported, where Optimus support WAN, LAN and MOCA.
Since the MindSpeed pfe_eth_init loops through
NUM_GEMAC_SUPPORT, which is set to 3, I leave the
comcerto_pfe_pdata array with also 3 elements, as in
Optimus. I continue to use "lan0" as the Ethernet port
name as the rest of the SpaceCast software initializes
the port using lan0 name. If people in SpaceCast agree,
we can change the name to more generic to eth1 as
recommended later. There will be another CL to fix the
atheros.c to properly set the PHY delay using the
board info.

Change-Id: I8e56389c1676c2dce7d49f9425a9d822f07bd3aa