Fix the Atheros PHY driver to add delay base on setting
Currently, the Atheros PHY driver adds delay to both RX
and TX in RGMII mode even though the setting is not to add.
Since the MindSpeed driver is fixed to return the correct
PHY delay setting based on the board config, the Atheros
driver is fixed to add appropriate delay to RX, TX, both or
none, based on interface setting.

Change-Id: I62c549a3d5cd03200d229b4784b08cd2efd99801