Correct BER calculation.
The current BER only displays the number of error bits, rather it should
be a ratio. In DVB-S, it should be error_bits/total_bits; in DVB-S2, it
should be error_frames/total/frames.

In linux DVB standard, there is no specification about the unit. Based
on the convention, 10^-9 is used as the unit since BER can only be set
as an 32 bit unsigned integer in DVB API.

Google-Bug-Id: 19734248
Change-Id: I7189535c3d8ac299d829baaf1ca0219810b4d2de