lm63: msleep(100) before retrying SMBus operation
Our experiments showed that the most common cause for failed SMBus
operations is bad code that blocks interrupts for too long. As a result,
the i2c driver fails to perform I2C operations in a timely manner.

If such a situation occurs, just sleep for 100ms hoping that the
situation has been resolved when we wake up and retry.

Change-Id: Ifd0e6e06caad0e406e603b53fc3e8ededfd552f8