i2c-comcerto: Limit max transaction time workaround to short transactions
The commit "i2c-comcerto: Limit duration of I2C transaction" (9d8683) limits
the maximum duration of an I2C transaction to address an issue with the lm96163
temperature sensor. This limitation causes issues with other chips like the
Infineon TPM, which perform transactions with larger amounts of data. With this
change the max duration of a transaction is only limited for transactions
shorter than 10 bytes.

Change-Id: I33d003fd42b46f1f378afbfc067a4432faef9e75