Add support for Comcerto 2000 to picoXcell driver.
The C2K has a crypto engine that is almost identical to picoXcell's
engine. Some registers have moved by an offset of 0x40000 so the
register defines have to be changed in the header.

The C2K supports AES-192 it does not need the software fallback
workaround for the picoXcell.

Change-Id: Ia6dce8f7666c5ce1a8cb09f9d5cecb0f77dc1a05