LS1024A NAND: Use HW ECC for reading, SW for writing
Due to a silicon bug where concurrent writes to NAND and NOR flash cause
NOR flash corruption when using the HW ECC engine for NAND, we went for
a hybrid approach where we use HW ECC only for reading and calculate the
ECC code in software when we are writing to NAND flash.

In the old 3.2 kernel, we used an ugly workaround where the NOR and NAND
driver shared a mutex to avoid concurrent access.

Change-Id: I4fcac8a5141d3ca6083e228d629239d8cce709e4