pinctrl driver for LS1024A
This is the first attempt at a pinctrl driver for the Freescale QorIQ
LS1024A SoC. It is complete enough to support our Google Fiber Optimus
board. Functions that are not used on our board have been left out.
Examples of functions that are not supported are low speed SPI, UART0
(we use UART1), CoreSight and TDM signals.
The GPIO driver has been merged into this pinctrl driver because it is
using the same register range and because of other dependencies.

Change-Id: I364c1f925c6d0c9fb15f184c802a4a20646fbb7a