Optimus: LEDs: increase max-brightness to 200
The target for maximum LED brightness as defined by the product
requirements is 50%. gpio-mailbox outputs "100" when it wants to set the
LED brightness to its maximum. Hence, we define 200 as the maximum LED
brightness in the device tree so that the LED operates at 50% when
gpio-mailbox sets the brightness to 100.

Change-Id: I8254e34be40918b6901451626006493b0d01fe44