Optimus: Switch on red LED by default
Set the default-trigger for the red LED to default-on. Otherwise, the
LED will stay dark after the pinctrl driver set the LED pins to PWM
until gpio-mailbox starts up and starts blinking the LED.

An extended period of LED darkness is confusing to the user. They might
think the box lost power. It would also be a deviation from the
behavior of the old kernel.

Change-Id: I1885ea815cc539396555bffbf89d537332ff07b2