dwc_otg: Only enable the ID pin interrupt in OTG mode
In certain circumstances (timing plays a critical role) dwc_otg ID pin
interrupt storms have been observed on gfsc100 at boot time. The
condition eventually leads to a corruption of the system clock, which
was the main symptom. The ID pin is only used in OTG mode to determine
whether a host or a device is connected to the USB port. Disable the ID
pin interrupt in host and device only mode.

Change-Id: Ice32ebac006a535da9015aa7a3501d001c26b708
Google-Bug-Id: 21036485