gfsc100: Disable CONFIG_COMCERTO_ZONE_DMA_NCNB
When CONFIG_COMCERTO_ZONE_DMA_NCNB is enabled, 44 MB of DRAM are mapped
non-cacheable and used to store IP packets that need to be forwarded
between wired Ethernet and Wifi. The advantage is that it isn't
necessary to run cache management operations like invalidate or
clean-and-invalidate a buffer is handed over to the hardware. These L2
cache operations (dma_map_single etc.) are pretty expensive. With this
architecture in place, it needs to be ensured that this type of memory
is not handed out to anything else because it would be very slow due to
the lack of the L2 caching. Mindspeed achieved this by declaring those
44 MB as ZONE_DMA and hacking the buddy allocator to not fall back to
ZONE_DMA if ZONE_NORMAL is exhausted.

SpaceCast does not perform packet forwarding and these 44 MB are in the
best case unused.

Google-Bug-Id: 24498038

Change-Id: I8de6095b9aac18d9ea637213a56d7e524a69ab4c