i2c-ls1024a: Remove feature to timeout I2C transactions
We previously had a feature where we would return an error if an I2C
transactions takes more than 22ms. This was useful with the LM96163
temperature sensor because it would reset itself if SDA or SCL are held
low for more than 35ms. Unfortunately, this feature caused trouble for
TPM chips that are connect over I2C.

The timeout feature has been moved to lm63.c. Remove it from the I2C
driver.

Change-Id: I0913e36a7dead52ba8b11dc6fbfd33f5a8a5f9e9