Designware SPI: Round up clock divisor
Previously, the clock divisor was rounded down which caused the clock
signal to oscillate at a higher frequency than what we specified. (20.83
MHz instead of 20 MHz). This is risky if the frequency we specified is
already the maximum for a certain component. Let's play it safe and
round up the divisor i.e. round down the frequency.
Unfortunately, the DesignWare SPI controller requires the divisor to be
even which can cause further frequency decrease.

Change-Id: I9fef4d18063122833acfb8d84c8950f27d98d82a