dwc_otg_hcd: Add delay before controller initialization
It has been observed that the controller doesn't work properly when the
kernel console is connected to a vterm (which affects timing).
Experiments have shown that the problem can be fixed with a delay of at
least 60ms before the controller initialization. A delay of 200ms is
used to add some margin.

Change-Id: I19f14898747bfc29cdad5b2806a69b5a8edebd50
Google-Bug-Id: 26006801