Add driver for sc100 DVB-S USB2.0 receiver
The hardware design is based on a Montage M88RS6000 tuner/demodulator,
a Cypress CY7C68013A USB bridge and a TI TPS65233 voltage regulator. The
M88RS6000 chip integrates a M88RS6000 tuner and a M88DS3103 demodulator.
The demodulator is directly connected to the I2C bus. The tuner is
connected internally to the M88DS3103 and can be accessed through an I2C
mux. The CY7C68013A USB bridge is connected to the MPEG-TS interface of
the M88RS6000 and makes the data on the MPEG-TS available to the host
through an USB bulk endpoint. The TPS65233 has an I2C interface, however
in the sc100 design it is controlled through pins driven by the
M88RS6000 chip.

Change-Id: I57a424c1c14472ba843da630d50ca93ac0ab8c41