dvbsky_m88rs6000: Report SNR in 0.1 dB
This is the convention followed by most DVB drivers. The upstream driver
for the m88ds3103 (demodulator part of the m88rs6000) reports SNR in 0.1
dB. It is planned to use this driver when moving to a 4.1 kernel. It is
preferable to make the dvbsky_m88rs6000 driver (and user space) follow
the convention rather than hacking the upstream driver to exhibit a
non-standard behavior.

Change-Id: Ida0966ec2d61733f690173212e856b2c4323015b