spi-dw: Enable slave only when tx fifo fully loaded
From the DesignWare DW_apb_ssi Databook:

"The Slave Enable Register (SER) register can be written here to enable
the target slave for selection. If a slave is enabled here, the transfer
begins as soon as one valid data entry is present in the transmit FIFO.
If no slaves are enabled prior to writing to the Data Register (DR), the
transfer does not begin until a slave is enabled."

"The transfer is stopped by the shift control logic when the transmit
FIFO is empty."

Previously, the driver set the SER (slave enable) before loading the TX
FIFO. The SPI hardware then started the transfer as soon as we wrote the
first data word into the data register which was the start of a race: If
the hardware transmits the first data word before we get a chance to
load subsequent words into the data register, the hardware would stop
the transfer and restart it as soon as we loaded the next word. Stopping
and restarting a transfer confuses the SPI slave and must be avoided.

Work around this issue by enabling the slave only after we fully loaded
the TX FIFO. This ensures that there is no stopping/restarting
mid transfer.

Change-Id: I84e061fbe5b89b70b93a16c841a5275fd3226474