gfsc100_defconfig: Build ext4, FAT and VFAT drivers as modules
The size of the gfsc100 4.1 kernel image is very close to exceed the
size of the flash partition. Removing the file system drivers from the
image gives us a margin of ~300kB. The drivers are only needed after the
squashfs root filesystem has been mounted and can be loaded at runtime.

Google-Bug-Id: 26519506

Change-Id: If39f1df9a6aa6e81624ae0541107d40018a8b8d3