m88ds3103: Fix BER calculation
The current BER only displays the number of error bits, rather it should
be a ratio. In DVB-S, it should be error_bits/total_bits; in DVB-S2,
it  should be error_frames/total_frames.

The Linux DVB standard doesn't specify a unit for the BER. Based on the
convention, 10^-9 is used as the unit since BER can only be set as a
32 bit unsigned integer in DVB API.

Ported from the 3.2 dvbsky driver

Google-Bug-Id: 19734248

Change-Id: I0bd98fccd3302a39df6e860f93a8fe90bb4dc322