LS1024A NAND: Fix up bit flips in empty ECC chunks
After an erase operation, all bits in a NAND page are set to 1. The ECC
engine has to special case empty ECC chunks (a page is split up into 4
ECC chunks) because the ECC code for all 1s is not all 1s. If it were to
take the ECC code literally, we would see an endless stream of
uncorrectable ECC errors.

The hardware ECC engine apparently does recognize empty ECC chunks, but
it wouldn't correct any bit flips therein. Ubifs reacts very
allergically to bit flips in pages that it previously erased: It prints
"ubifs_recover_leb: corrupt empty space" and refuses to mount the
volume. Experiments showed that the hardware engine defines an empty
page as a page with less than 12 bits set to one.

Change-Id: Icde5005d5d061a238eb96128c37114c1df220792