tpm_i2c_infineon: Add reboot handler
On gfsc100 the I2C bus sometimes does not work at all after a reboot.
The problem happens when the reboot occurs in midst of an I2C read
transfer from the TPM. The TPM is not physically reset on reboots and
apparently remains in a state where it keeps holding the bus to complete
the pending read transfer.

To avoid this situation a reboot handler is added to the TPM driver. The
handler sets a flag indicating that a reboot is underway and flushes the
work queue of the I2C controller. Subsequent attempts to read or write
are aborted.

Google-Bug-Id: 23816953

Change-Id: I58c5c6cd4485315b2850f731e1fb28a53d13cac4