Add tcp acubic, transonic, and turbulent congestion control options.
See Config.in for explanations of each.

We can use these for local video streams over wifi.  In medium-to-high
latency situations with a bit of packet loss, transonic seems to get about
twice the throughput capacity as cubic, reno, or westwood.

Example command line to run (on the sender) to simulate poor conditions:
  tc qdisc add dev wlan0 root netem limit 1000 loss 1% delay 30ms

Note that these algorithms are not a good choice as the default TCP
congestion control.  They only work reasonably for fixed-rate streams (such
as video streaming) on local area networks.  When used to stream large
amounts of data, they might flood out any competitors on the link.

Also note:

- Because of the way we set the congestion window, transonic and turbulent
  work much better with packet loss than with variable latency.  We might
  want to disable hardware-layer automatic retransmits and let TCP do it.

- Once the congestion window is held open, the next likely bottleneck will
  be the rwnd on the receiver, and/or the send buffer on the transmitter.
  The default rwin settings in /proc/sys/net/{core,ipv4} are way too small.
  (Increasing them also improves performance of cubic/reno/westwood on poor
  quality links.)

- On the receiver:
  echo 0 >/proc/sys/net/ipv4/tcp_moderate_rcvbuf

- On the transmitter:
  echo 1 >/proc/sys/net/ipv4/tcp_thin_linear_timeouts
  echo 1 >/proc/sys/net/ipv4/tcp_thin_dupack
  echo 10 >/proc/sys/net/ipv4/tcp_min_rtt_wlen
  echo 4 >/proc/sys/net/ipv4/tcp_early_retrans  # helped a lot
  echo 0 >/proc/sys/net/ipv4/tcp_reordering     # maybe

These experiments are inspired by the work of the Google TCP team,
especially Nandita Dukkipati, Eric Dumazet, and Stephen Gunn.

Change-Id: If3ca1025b147cd9cf295d3c81f8c8b902ba2baaa