br: Add all local unicast addresses to all slaves
For the fast forwarding feature in the NXP QorIQ LS1024A SoC (formerly
Mindspeed Comcerto 2000) to work, we not only need to set the Ethernet
interfaces to promiscuous mode, we also have to program all local MAC addresses
of the br0 interface into all slave interfaces. The set of local MAC
addresses usually consists of the MAC addresses of all slave devices and
the MAC address of the bridge interface. The latter is usually identical
to one of the slave interfaces.

We deliberately do not avoid adding an interface's own MAC address via a
call to dev_uc_add() which results in a redundant MAC address
assignment. This decision has been made to avoid complexity.

Change-Id: I2633139046a0922d1d171d8d30316a3142c83324