printk: Flush incomplete log messages before reboot
The following pr_emerg() statement in kernel/panic.c is missing a line
feed character as the last character in the format string:

pr_emerg("Rebooting in %d seconds..", panic_timeout);

As a result, the message is not immediately stored in the log buffer and
is subsequently lost unless another printk() statement is executed prior
to the reboot.

Let's flush these incomplete log lines prior to reboot to ensure that
the string

"Rebooting in 3 seconds.."

is preserved as the last log message in the buffer. Otherwise, a user
might assume that printk_persist lost the last few message prior to
reboot.

Change-Id: I2528de7fd297a8ce9186b8a2ffe18b365773a9e9