drm/meson: convert to the new canvas module
This removes the meson_canvas files within the meson/drm layer
and makes use of the new canvas module that is referenced in the dts.

Canvases can be used by different IPs and modules, and it is as such
preferable to rely on a module that can safely dispatch canvases on
demand.

Signed-off-by: Maxime Jourdan <maxi.jourdan@wanadoo.fr>

platform 665dbad80386dcb9ac78ec6153d8d518da9bb50e